+++ test/coverage_gap_2_test.dart
@@ -61,7 +61,8 @@ class _FakeErrorHandler extends Fake implements ErrorInterceptorHandler {
   void resolve(Response<dynamic> response) => onResolve?.call(response);
 
   @override
-  void reject(DioException err) => onReject?.call(err);
+  void reject(DioException err, [bool callFollowingErrorInterceptor = false]) =>
+      onReject?.call(err);
 }
 
 class _FakeRequestHandler extends Fake implements RequestInterceptorHandler {